+++ src/unreal/structs/gobjects.cpp
@@ -7,7 +7,7 @@ namespace unrealsdk::unreal {
 #ifdef UE4
 
 FUObjectItem* FChunkedFixedUObjectArray::at(size_t idx) const {
-    if (idx > this->Count) {
+    if (idx > (size_t)this->Count) {
         throw std::out_of_range("FChunkedFixedUObjectArray index out of range");
     }
     // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
