+++ src/unreal/wrappers/wrapped_struct.cpp
@@ -11,7 +11,7 @@ namespace unrealsdk::unreal {
 void copy_struct(uintptr_t dest, const WrappedStruct& src) {
     for (const auto& prop : src.type->properties()) {
         cast_prop(prop, [dest, &src]<typename T>(const T* prop) {
-            for (size_t i = 0; i < prop->ArrayDim; i++) {
+            for (size_t i = 0; i < (size_t)prop->ArrayDim; i++) {
                 set_property<T>(prop, i, dest, src.get<T>(prop, i));
             }
         });
@@ -21,7 +21,7 @@ void copy_struct(uintptr_t dest, const WrappedStruct& src) {
 void destroy_struct(const UStruct* type, uintptr_t addr) {
     for (const auto& prop : type->properties()) {
         cast_prop(prop, [addr]<typename T>(const T* prop) {
-            for (size_t i = 0; i < prop->ArrayDim; i++) {
+            for (size_t i = 0; i < (size_t)prop->ArrayDim; i++) {
                 destroy_property<T>(prop, i, addr);
             }
         });
